+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Version 1.3.0+9 R.C
 
+ - Calling disconnect if state could not be retrieved.
  - (Linux) Added basic support for a linux version.
  - (Web) Fixed bug where devices would show inside the group and outside the group.
  - Vive base station ids are now bound to the device id instead to the device name.